Warning fix, intentional cast

+++ src/core/engine/rendering/lowlevelapi/VertexBufferLayout.cpp
@@ -4,6 +4,6 @@
 
 void VertexBufferLayout::push(unsigned int type, unsigned int count, bool normalized) {
     size_t typeSize = GLsizeof(type);
-    m_elements.push_back({type, typeSize, count, normalized});
+    m_elements.push_back({type, static_cast<unsigned int>(typeSize), count, normalized});
     m_stride += count * typeSize;
 }
